+++ Changelog.md
@@ -1,6 +1,7 @@
 ### 0.4.10 (unreleased)
 
 Features:
+ * Commandline interface: Do not overwrite files unless forced.
  * Add ``assert(condition)``, which throws if condition is false.
  * Introduce ``.transfer(value)`` for sending Ether.
  * Code generator: Support ``revert()`` to abort with rolling back, but not consuming all gas.